20130602 home combine
Add sales activity self.report
Add logicaldelete and afterwards add to sales activity record
Add new style to show logical delete and afterwards add in record,
self.list, manager.list and global.list

To add overtime application

+++ modules/sales/activity.php
@@ -9,10 +9,18 @@ switch($A){
 	            OR owner_employee_id LIKE "'.$myId.',%" 
 	            OR owner_employee_id LIKE "%,'.$myId.',%") ORDER BY name ASC, id ASC');
 	    $customer = basicMysqliQuery('uatme_oa_sales_customer', ' WHERE ifavailable="1"');
-		$minday = Date('Y-m-d',(time()-Date('w')*24*3600));
-		$maxday = Date('Y-m-d',(time()+(6-Date('w'))*24*3600));
+		$minday = Date('Y-m-d',(time()-(Date('w')+7)*24*3600));
+		$maxday = Date('Y-m-d',(time()+(13-Date('w'))*24*3600));
 	    $activity = basicMysqliQuery('uatme_oa_sales_activity', ' WHERE employee_id="'.$_SESSION["employee_id"].'" AND date BETWEEN "'.$minday.'" AND "'.$maxday.'"');
+	    
+	    $deletableminday = Date('Y-m-d',(time()-(Date('w', $selectedUTCStamp)+7)*24*3600));
+	    $deletablemaxday = Date('Y-m-d',(time()+(13-Date('w', $selectedUTCStamp))*24*3600));
 	    foreach($activity as $a){
+	        if($a['date'] >= $deletableminday and $a['date'] <= $deletablemaxday and $a['iflogicaldeleted']==0){
+	            $a['deletable'] = 1;
+	        }else{
+	            $a['deletable'] = 0;
+	        }
 	        $a['customerName'] = $customer[$a['customer_id']]['name'];
 	        $assign['activity'][] = $a;
 	    }
@@ -32,17 +40,28 @@ switch($A){
 		}
 		if($customerId > 0){
     		$detail = $_POST['detail'];
-    		$minday = Date('Y-m-d',(time()-Date('w')*24*3600));
-    		$maxday = Date('Y-m-d',(time()+(6-Date('w'))*24*3600));
-    		if($date >= $minday and $date <= $maxday){
-    		    $sql = 'INSERT INTO uatme_oa_sales_activity(date, start, end, customer_id, detail, employee_id)
-    		             VALUES ("'.$date.'", "'.$start.'", "'.$end.'", "'.$customerId.'", "'.$detail.'", "'.$_SESSION["employee_id"].'")';
+    		//add event for last week
+    		$minday1 = Date('Y-m-d',(time()-(Date('w')+7)*24*3600));
+    		$maxday1 = Date('Y-m-d',(time()-(Date('w')+1)*24*3600));
+    		if($date >= $minday1 and $date <= $maxday1){
+    		    $sql = 'INSERT INTO uatme_oa_sales_activity(date, start, end, customer_id, detail, employee_id, ifafterwards)
+    		             VALUES ("'.$date.'", "'.$start.'", "'.$end.'", "'.$customerId.'", "'.$detail.'", "'.$_SESSION["employee_id"].'", 1)';
     		    $mysqli->query($sql);
     		    $httpstatus = 200;
-    		    $msg = '活动记录成功！';
-    		}else{
-    		    $httpstatus = 500;
-    		    $error = '记录日期 ('.$date.') 超出允许范围 ('.$minday.'~'.$maxday.')';
+    		    $msg = '活动记录成功！';  		    
+    		}else{	
+        		$minday2 = Date('Y-m-d',(time()-Date('w')*24*3600));
+        		$maxday2 = Date('Y-m-d',(time()+(13-Date('w'))*24*3600));
+        		if($date >= $minday2 and $date <= $maxday2){
+        		    $sql = 'INSERT INTO uatme_oa_sales_activity(date, start, end, customer_id, detail, employee_id)
+        		             VALUES ("'.$date.'", "'.$start.'", "'.$end.'", "'.$customerId.'", "'.$detail.'", "'.$_SESSION["employee_id"].'")';
+        		    $mysqli->query($sql);
+        		    $httpstatus = 200;
+        		    $msg = '活动记录成功！';
+        		}else{
+        		    $httpstatus = 500;
+        		    $error = '记录日期 ('.$date.') 超出允许范围 ('.$minday1.'~'.$maxday2.')';
+        		}
     		}
 		}else{
 		    $httpstatus = 500;
@@ -52,10 +71,29 @@ switch($A){
 		break;
 	case 'delete':
 		$id = $_POST['id'];
-		$sql = 'DELETE FROM uatme_oa_sales_activity WHERE id="'.$id.'"';
-		$mysqli->query($sql);
-		$httpstatus = 200;
-		$msg = '记录成功删除';
+		$activity = basicMysqliQuery('uatme_oa_sales_activity', ' WHERE id="'.$id.'" ');
+		//logical delete event for last week
+		$minday1 = Date('Y-m-d',(time()-(Date('w')+7)*24*3600));
+		$maxday1 = Date('Y-m-d',(time()-(Date('w')+1)*24*3600));
+		if($activity[$id]['date'] >= $minday1 and $activity[$id]['date'] <= $maxday1){
+    		$sql = 'UPDATE uatme_oa_sales_activity SET iflogicaldeleted="1" WHERE id="'.$id.'"';
+    		$mysqli->query($sql);
+    		$httpstatus = 200;
+    		$msg = '记录删除标记成功';
+		}else{
+		    //real delete
+		    $minday2 = Date('Y-m-d',(time()-Date('w')*24*3600));
+		    $maxday2 = Date('Y-m-d',(time()+(13-Date('w'))*24*3600));
+		    if($activity[$id]['date'] >= $minday2 and $activity[$id]['date'] <= $maxday2){
+        		$sql = 'DELETE FROM uatme_oa_sales_activity WHERE id="'.$id.'"';
+        		$mysqli->query($sql);
+        		$httpstatus = 200;
+        		$msg = '记录成功删除';
+		    }else{
+    		    $httpstatus = 500;
+    		    $error = '选定的记录日期 ('.$date.') 超出允许删除范围 ('.$minday1.'~'.$maxday2.')';
+		    }
+		}
 		sendResponse($httpstatus, $error, $msg);
 		break;
 	case 'manager.list':
@@ -69,7 +107,7 @@ switch($A){
 	    $selectedUTCStamp = mktime(0,0,0,$selectedMonth, $selectedDay, $selectedYear); 
 		$minday = Date('Y-m-d',($selectedUTCStamp-Date('w', $selectedUTCStamp)*24*3600));
 		$maxday = Date('Y-m-d',($selectedUTCStamp+(6-Date('w', $selectedUTCStamp))*24*3600));
-	    $wherePeriod = ' AND (createdate BETWEEN "'.$minday.'" AND "'.$maxday.'") ';
+	    $wherePeriod = ' AND (date BETWEEN "'.$minday.'" AND "'.$maxday.'") ';
 	    $assign['selectedPeriod'] = $minday . ' ~ ' . $maxday;
 	    //echo $wherePeriod;    
 	    
@@ -81,7 +119,7 @@ switch($A){
 	    foreach($myEmployeeId as $d=>$e){
 	        $activity[$d]['department_name'] = $department[$d]['name'];
 	        $activity[$d]['department_id'] = $d;
-	        $activity[$d]['activity'] = basicMysqliQuery('uatme_oa_sales_activity', ' WHERE employee_id IN ('.implode(',',$e).') '.$wherePeriod.' ORDER BY employee_id ASC, createdate DESC');
+	        $activity[$d]['activity'] = basicMysqliQuery('uatme_oa_sales_activity', ' WHERE employee_id IN ('.implode(',',$e).') '.$wherePeriod.' ORDER BY employee_id ASC, date DESC');
 	    }
 	    //print_r($activity);
 	    $assign['activity'] = $activity;
@@ -99,14 +137,14 @@ switch($A){
 	    $selectedUTCStamp = mktime(0,0,0,$selectedMonth, $selectedDay, $selectedYear); 
 		$minday = Date('Y-m-d',($selectedUTCStamp-Date('w', $selectedUTCStamp)*24*3600));
 		$maxday = Date('Y-m-d',($selectedUTCStamp+(6-Date('w', $selectedUTCStamp))*24*3600));
-	    $wherePeriod = ' AND (createdate BETWEEN "'.$minday.'" AND "'.$maxday.'") ';
+	    $wherePeriod = ' AND (date BETWEEN "'.$minday.'" AND "'.$maxday.'") ';
 	    $assign['selectedPeriod'] = $minday . ' ~ ' . $maxday;
 	    //echo $wherePeriod;	
 	    
 	    $department = basicMysqliQuery('uatme_oa_system_department');
 	    $assign['customer'] = basicMysqliQuery('uatme_oa_sales_customer', ' WHERE ifavailable="1"');
 	    $assign['employee'] = basicMysqliQuery('uatme_oa_system_employee', ' WHERE ifavailable="1"');
-	    $activity = basicMysqliQuery('uatme_oa_sales_activity',  ' WHERE 1'.$wherePeriod.' ORDER BY employee_id ASC, createdate DESC');
+	    $activity = basicMysqliQuery('uatme_oa_sales_activity',  ' WHERE 1'.$wherePeriod.' ORDER BY employee_id ASC, date DESC');
 	    foreach($activity as $a){
 	        $department_id = $assign['employee'][$a['employee_id']]['department_id'];
 	        $assign['activity'][$department_id]['department_name'] = $department[$department_id]['name'];
@@ -128,13 +166,24 @@ switch($A){
 	    $selectedUTCStamp = mktime(0,0,0,$selectedMonth, $selectedDay, $selectedYear); 
 		$minday = Date('Y-m-d',($selectedUTCStamp-Date('w', $selectedUTCStamp)*24*3600));
 		$maxday = Date('Y-m-d',($selectedUTCStamp+(6-Date('w', $selectedUTCStamp))*24*3600));
-	    $wherePeriod = ' AND (createdate BETWEEN "'.$minday.'" AND "'.$maxday.'") ';
+	    $wherePeriod = ' AND (date BETWEEN "'.$minday.'" AND "'.$maxday.'") ';
 	    $assign['selectedPeriod'] = $minday . ' ~ ' . $maxday;
 	    //echo $wherePeriod;	
 	    
 	    $department = basicMysqliQuery('uatme_oa_system_department');
 	    $assign['customer'] = basicMysqliQuery('uatme_oa_sales_customer', ' WHERE ifavailable="1"');
-	    $assign['activity'] = basicMysqliQuery('uatme_oa_sales_activity',  ' WHERE owner_employee_id="'.$_SESSION['employee_id'].'" '.$wherePeriod.' ORDER BY createdate DESC');
+	    $assign['activity'] = basicMysqliQuery('uatme_oa_sales_activity', ' WHERE employee_id="'.$_SESSION['employee_id'].'" '.$wherePeriod.' ORDER BY date DESC');
+	    
+	    //mark which activity are deleteable
+	    $deletableminday = Date('Y-m-d',(time()-(Date('w', $selectedUTCStamp)+7)*24*3600));
+	    $deletablemaxday = Date('Y-m-d',(time()+(13-Date('w', $selectedUTCStamp))*24*3600));
+	    foreach($assign['activity'] as $a){
+	        if($a['date'] >= $deletableminday and $a['date'] <= $deletablemaxday and $a['iflogicaldeleted']==0){
+	            $assign['activity'][$a['id']]['deletable'] = 1;
+	        }else{
+	            $assign['activity'][$a['id']]['deletable'] = 0;
+	        }
+	    }
 	    
 	    //print_r($activity);
 		$smarty->assign($assign);
